Local pong (#20)
* merge main into valmpani

Signed-off-by: valmpani <vasilisalmpanis@gmail.com>

* transcendence: backend: Fixed small bug with game types

Signed-off-by: valmpani <vasilisalmpanis@gmail.com>

* transcendence: frontend: fix: local games

Adds local games to frontend and fixes collisions
with paddle to work like the classic game pong where
the further away the ball is from the center of the
paddle the bigger the bounce angle

Signed-off-by: valmpani <vasilisalmpanis@gmail.com>

* transcendece: backend: fix: better collisions

Signed-off-by: valmpani <vasilisalmpanis@gmail.com>

* transcendence: cli: Error messages and handling of game state

Fixes all situations where 401 might arise and client is notified
so they can authorize again. Better error messages separated and
predined so they can be easily changed. Adds type hints on all
functions and some descriptions to the most important functions.
Adds script for automatization of package building and running.

Signed-off-by: Vasileios Almpanis <vasilisalmpanis@gmail.com>

* sleep in game runner for smoother experience

* removes logger messages

* fixes infinite refresh from apiClient and some components

* alert component in profile

---------

Signed-off-by: valmpani <vasilisalmpanis@gmail.com>
Signed-off-by: Vasileios Almpanis <vasilisalmpanis@gmail.com>